feat(census-models): Add new model buttons for copy/navigate (#6343)
