+++ frontend/census-projects.json
@@ -21,7 +21,12 @@
     "n_cells": 36227903,
     "n_columns": 512,
     "n_features": 512,
-    "notebook_links": [["Using trained model", "https://chanzuckerberg.github.io/cellxgene-census/notebooks/analysis_demo/comp_bio_geneformer_prediction.html"]],
+    "notebook_links": [
+      [
+        "Using trained model",
+        "https://chanzuckerberg.github.io/cellxgene-census/notebooks/analysis_demo/comp_bio_geneformer_prediction.html"
+      ]
+    ],
     "submission_date": "2023-11-06",
     "last_updated": null,
     "revised_by": null
@@ -65,7 +70,12 @@
     "submission_date": "2023-11-18",
     "publication_info": "",
     "publication_link": "",
-    "notebook_links": [["Using trained model", "https://chanzuckerberg.github.io/cellxgene-census/notebooks/analysis_demo/comp_bio_scvi_model_use.html"]],
+    "notebook_links": [
+      [
+        "Using trained model",
+        "https://chanzuckerberg.github.io/cellxgene-census/notebooks/analysis_demo/comp_bio_scvi_model_use.html"
+      ]
+    ],
     "last_updated": null,
     "revised_by": null
   },
@@ -108,7 +118,12 @@
     "submission_date": "2023-11-18",
     "publication_info": "",
     "publication_link": "",
-    "notebook_links": [["Using trained model", "https://chanzuckerberg.github.io/cellxgene-census/notebooks/analysis_demo/comp_bio_scvi_model_use.html"]],
+    "notebook_links": [
+      [
+        "Using trained model",
+        "https://chanzuckerberg.github.io/cellxgene-census/notebooks/analysis_demo/comp_bio_scvi_model_use.html"
+      ]
+    ],
     "last_updated": null,
     "revised_by": null
   }
