+++ frontend/src/common/analytics/events.ts
@@ -53,6 +53,7 @@ export enum EVENTS {
   CENSUS_DIRECTORY_CLICK_NAV = "CENSUS_DIRECTORY_CLICK_NAV",
   CENSUS_EMBEDDING_CLICKED = "CENSUS_EMBEDDING_CLICKED",
   CENSUS_MODEL_CLICKED = "CENSUS_MODEL_CLICKED",
+  CENSUS_MODEL_COPIED = "CENSUS_MODEL_COPIED",
   CENSUS_PUBLICATION_CLICKED = "CENSUS_PUBLICATION_CLICKED",
   CENSUS_CONTACT_CLICKED = "CENSUS_CONTACT_CLICKED",
   CENSUS_NOTEBOOK_CLICKED = "CENSUS_NOTEBOOK_CLICKED",
